requirements.txt
python-dotenv
msal
requests
pandas
openpyxl
groq
flask
apscheduler
pytz
tzlocal